+++ central-portal/src/main/kotlin/com/vanniktech/maven/publish/portal/SonatypeCentralPortalService.kt
@@ -3,11 +3,11 @@ package com.vanniktech.maven.publish.portal
 import okhttp3.MultipartBody
 import okhttp3.ResponseBody
 import retrofit2.Call
-import retrofit2.http.Body
 import retrofit2.http.DELETE
 import retrofit2.http.GET
 import retrofit2.http.Multipart
 import retrofit2.http.POST
+import retrofit2.http.Part
 import retrofit2.http.Path
 import retrofit2.http.Query
 import retrofit2.http.Streaming
@@ -50,7 +50,7 @@ internal interface SonatypeCentralPortalService {
   fun uploadBundle(
     @Query("name") name: String?,
     @Query("publishingType") publishingType: String?,
-    @Body input: MultipartBody.Part,
+    @Part input: MultipartBody.Part,
   ): Call<String>
 
   @Streaming